Add files via upload
The proposed project titled under UCR is “10 MW Toss Small Hydro Electric Project in Kullu District of Himachal Pradesh”, which is a Hydro Power project located in village Toss of Kullu district in the state of Himachal Pradesh (India). The project is an operational activity with continuous reduction of GHG, currently being applied for registration under “Universal Carbon Registry” (UCR).